test: run daemon file-API tests against the real workspace-root provider
Two homeboy-core daemon file-API tests resolve a request path against a runner's
configured workspace root, which the daemon `/files/*` routes look up through the
`RunnerWorkspaceRootProvider` hook. That provider's real implementation lives in
the runner crate (extracted from core in #8698), so in homeboy-core these tests
hit the no-op provider and fail closed with "requires workspace_root" — and a
core-only stub would just be a test double.

Move them to homeboy-lab-runner, which registers the real provider
(`register_runner_workspace_root_provider`), so path containment is enforced
against the runner's actual configured root:

- `file_route_rejects_paths_outside_runner_workspace_root`
- `file_routes_upload_and_download_inside_runner_workspace_root`

They dispatch in process through the public `homeboy_core::daemon::route_with_body`
router (added when the /exec result tests moved over). Both pass in lab-runner
with the real provider. The core file-API tests that don't need the provider —
`file_route_accepts_request_workspace_root_without_runner_config` (path carried in
the request) and `file_route_requires_broker_submit_auth_for_untrusted_requests`
(auth-denied before workspace resolution) — stay in core and still pass.
